Add retention option and new resource types in Jira provider
Added a new attribute "RetainOnDelete" to the Jira Project resource which allows control over whether the project is deleted when the Terraform resource is destroyed. Furthermore, introduced new resource types: `ConfluenceSpaceResource` and `BoardResource`. These changes also included corresponding modifications to the `ProjectModel` and methods related to resource creation and deletion.